refactor(ui): split cart and payment large files into helpers
