+++ FakeNFT/Scenes /Cart/Notification+Cart.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+extension Notification.Name {
+    static let cartDidChange = Notification.Name("cartDidChange")
+}
